+++ shenyu-admin-listener/pom.xml
@@ -36,4 +36,5 @@
         <module>shenyu-admin-listener-polaris</module>
         <module>shenyu-admin-listener-zookeeper</module>
     </modules>
+
 </project>
