+++ shenyu-admin-listener/shenyu-admin-listener-etcd/pom.xml
@@ -33,19 +33,14 @@
         </dependency>
 
         <dependency>
-            <groupId>io.etcd</groupId>
-            <artifactId>jetcd-core</artifactId>
-        </dependency>
-
-        <dependency>
             <groupId>org.apache.shenyu</groupId>
-            <artifactId>shenyu-admin-listener-api</artifactId>
+            <artifactId>shenyu-infra-etcd</artifactId>
             <version>${project.version}</version>
         </dependency>
 
         <dependency>
             <groupId>org.apache.shenyu</groupId>
-            <artifactId>shenyu-common</artifactId>
+            <artifactId>shenyu-admin-listener-api</artifactId>
             <version>${project.version}</version>
         </dependency>
