+++ shenyu-admin-listener/shenyu-admin-listener-etcd/src/main/java/org/apache/shenyu/admin/listener/etcd/EtcdDataDataChangedListener.java
@@ -19,6 +19,7 @@ package org.apache.shenyu.admin.listener.etcd;
 
 import org.apache.shenyu.admin.listener.AbstractPathDataChangedListener;
 import org.apache.shenyu.common.utils.GsonUtils;
+import org.apache.shenyu.infra.etcd.EtcdClient;
 
 /**
  * EtcdDataDataChangedListener.
