+++ shenyu-infra/shenyu-infra-etcd/src/main/java/org/apache/shenyu/infra/etcd/EtcdConstants.java
@@ -1,0 +1,11 @@
+package org.apache.shenyu.infra.etcd;
+
+/**
+ * Etcd constants define.
+ */
+
+public interface EtcdConstants {
+
+    String ETCD = "etcd";
+
+}
