+++ shenyu-infra/shenyu-infra-etcd/src/main/java/org/apache/shenyu/infra/etcd/EtcdProperties.java
@@ -1,0 +1,4 @@
+package org.apache.shenyu.infra.etcd;
+
+public class EtcdProperties {
+}
