+++ shenyu-registry/shenyu-registry-etcd/pom.xml
@@ -33,6 +33,12 @@
         </dependency>
 
         <dependency>
+            <groupId>org.apache.shenyu</groupId>
+            <artifactId>shenyu-infra-etcd</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
+        <dependency>
             <groupId>io.etcd</groupId>
             <artifactId>jetcd-core</artifactId>
             <exclusions>
@@ -63,22 +69,6 @@
             </exclusions>
         </dependency>
 
-        <dependency>
-            <groupId>io.grpc</groupId>
-            <artifactId>grpc-core</artifactId>
-            <version>${grpc.version}</version>
-        </dependency>
 
-        <dependency>
-            <artifactId>grpc-protobuf</artifactId>
-            <groupId>io.grpc</groupId>
-            <version>${grpc.version}</version>
-        </dependency>
-
-        <dependency>
-            <artifactId>grpc-stub</artifactId>
-            <groupId>io.grpc</groupId>
-            <version>${grpc.version}</version>
-        </dependency>
     </dependencies>
 </project>
