+++ shenyu-spring-boot-starter/shenyu-spring-boot-starter-sync-data-center/shenyu-spring-boot-starter-sync-data-etcd/pom.xml
@@ -33,4 +33,5 @@
             <version>${project.version}</version>
         </dependency>
     </dependencies>
+
 </project>
